Add patients to a subset

Add patients to a subset, only if not already in the subset

add_patients_to_subset(
  output,
  r = shiny::reactiveValues(),
  patients = tibble::tibble(),
  subset_id = integer(),
  success_notification = TRUE,
  language = "EN"
)

Arguments

output
Output variable from Shiny, used to render messages on the message bar

r
The r reactive variable, used to communicate between modules in the ShinyApp (reactiveValues object)

patients
data variable containing patients (data.frame / tibble)

subset_id
ID of subset (integer)

success_notification
Should a message bar be displayed if insertion of patient(s) is a success ? Default to TRUE

language
language used for error / warning messages (character, default = "EN")

Details

...

Examples

if (FALSE) {
patients <- tibble::tribble(~patient_id, 123L, 456L, 789L)
subset_add_patients(output = output, r = r, patients = patients, subset_id = 3, language = "EN")
}