Run the Shiny Application

Runs the cdwtools Shiny Application.
Use language argument to choose language to use ("EN" or "FR" available).
Use options to set options to shiny::shinyApp() function, see documentation of shiny::shinyApp for more informations.

To connect to distant database, use db_info argument. db_info is a list, containing these parameters :

sql_lib : SQL library used, "postgres" or "sqlite" available for now

dbname : database name (character)

host : host connection name (character)

port : connection port (integer)

user : user name (character)

password : password in clear (character)

cdwtools(
  language = "EN",
  db_info = list(),
  default_folder = character(),
  options = list(),
  ...
)

Arguments

language
Default language to use in the App (character)

db_info
Database connection informations, if it is needed to connect a distant db (list).

default_folder
Default folder where to save datamarts CSV files (character)

options
A list of options that could be passed to shiny::shinyApp() function (list)

...
arguments to pass to golem_opts.

Examples