Create a datamart

Loads +/- save data when a datamart is chosen by the user.

create_datamart(
  r = shiny::reactiveValues(),
  datamart_id = integer(),
  data = tibble::tibble(),
  type = "patients",
  save_as_csv = TRUE,
  rewrite = FALSE,
  language = "EN"
)

Arguments

r
The "petit r" object, used to communicate between modules in the ShinyApp (reactiveValues object)

datamart_id
ID of the datamart, used to create directory in data/ (eg : datamart_3)

data
data variable (data.frame or tibble)

type
type or name of data to create c("patients", "stays", "labs_vitals", "text", "orders", "thesaurus")

save_as_csv
save or not the data to CSV file (logical, default = TRUE)

rewrite
if save_as_csv is TRUE, rewrite or not existing CSV file (logical, default = FALSE)

language
language used for error / warning messages (character, default = "EN")

Details

The function is used in a datamart code and is launched each time a user selects a datamart.
You can choose to load data each time the function is used with save_as_csv set to FALSE (eg when datamart is small and the connection to source database is good) or you can save the data in a CSV file with save_as_csv set to TRUE.
Basically, 6 data variables are created for each datamart (distinct values of 'type' parameter).

Examples

if (FALSE) {
patients <- tibble::tribble(~patient_id, ~gender, ~age, ~dod, 44565L, "F", 45, "2021-05-01 00:00:00") %>%
    dplyr::mutate_at("dod", lubridate::ymd_hms)
    
create_datamart(r = r, datamart_id = 5, data = patients, type = "patients", 
    save_as_csv = FALSE, rewrite = FALSE, language = "EN")
}