Render UI of settings creation card

Render UI of settings creation card

render_settings_creation_card(
  language = "EN",
  ns = shiny::NS(),
  id = character(),
  title = character(),
  textfields = character(),
  textfields_width = "200px",
  dropdowns = character(),
  dropdowns_width = "200px",
  words = tibble::tibble()
)

Arguments

language

Language used (character)

ns

Shiny namespace

id

ID of current module / page (character)

title

Title used to create the card, it will be translated with translate function (character)

textfields

A character vector containing distinct textfields to render in the card (character)

textfields_width

Width of the textfields, CSS code, so it could be "100%", "200px" etc (character)

dropdowns

A character vector containing distinct dropdowns to render in the card (character)

dropdowns_width

Width of the dropdowns, CSS code, so it could be "100%", "200px" etc (character)

Value

Shiny UI elements / HTML code

Examples

render_settings_creation_card(language = "EN", ns = NS("settings_datamart"), title = "create_datamart",
textfields = c("name", "description"), dropdowns = "data_source")
#> Error in render_settings_creation_card(language = "EN", ns = NS("settings_datamart"),     title = "create_datamart", textfields = c("name", "description"),     dropdowns = "data_source"): could not find function "render_settings_creation_card"